add edit oper_klient + fix kosyaki

+++ shpagin_podolsky/oper_klient/view.php
@@ -10,10 +10,6 @@
         require_once '../header.php';
 
         $tableRows = $mysqli -> query("SELECT id_operklient,number_account, name_operation, date_operation, price_operation FROM `schet`, `s_operation`, `oper_klient` WHERE oper_klient.id_schet = schet.id_schet AND s_operation.id_operation=oper_klient.id_operation");
-
-        function isDisabled($id){
-            
-        }
     ?>
     <div class="container text-center">
         <table class="table dict">
@@ -29,13 +25,13 @@
             <tbody>
                 <?php while($row = $tableRows -> fetch_assoc()){ extract($row) ?>
                 <tr>
-                    <td ><?= $row['number_account']?></td>
-                    <td ><?= $row['name_operation']?></td>
-                    <td ><?= $row['date_operation']?></td>
-                    <td ><?= $row['price_operation']?></td>
+                    <td ><?= $number_account?></td>
+                    <td ><?= $name_operation?></td>
+                    <td ><?= $date_operation?></td>
+                    <td ><?= $price_operation?></td>
                     <td class="btns"> 
                         <a type="button" class="btn btn-secondary" href="edit.php?id=<?=$id_operklient?>">Редактировать</a> 
-                        <a type="button" class="btn btn-secondary" href="delete.php?id=<?=$id_operklient?>">Удалить</a>
+                        <a type="button" class="btn btn-secondary" href="delete.php?id=<?=$id_operklient?>" onclick='return confirm(`Вы действительно хотите удалить запись "<?="$number_account - $name_operation - $date_operation"?>"?`);'>Удалить</a>
                     </td>
                 </tr>
                 <?php } ?>
