+++ shpagin_podolsky/sheta/view.php
@@ -9,10 +9,6 @@
         require_once 'utils.php';
         
         $tableRows = $mysqli -> query("SELECT id_schet ,name_klient, name_account, number_account, date_open, date_close, price_open FROM `schet`, `s_accoount`, `klient` WHERE klient.id_klient = schet.id_klient AND schet.id_account=s_accoount.id_account");
-
-        function isDisabled($id){
-            
-        }
     ?>
     <div class="container text-center">
         <table class="table dict">
@@ -30,15 +26,15 @@
             <tbody>
                 <?php while($row = $tableRows -> fetch_assoc()){ extract($row)?>
                 <tr>
-                    <td ><?= $row['name_klient']?></td>
-                    <td ><?= $row['name_account']?></td>
-                    <td ><?= $row['number_account']?></td>
-                    <td ><?= $row['date_open']?></td>
-                    <td ><?= $row['date_close']?></td>
-                    <td ><?= $row['price_open']?></td>
+                    <td ><?= $name_klient?></td>
+                    <td ><?= $name_account?></td>
+                    <td ><?= $number_account?></td>
+                    <td ><?= $date_open?></td>
+                    <td ><?= $date_close?></td>
+                    <td ><?= $price_open?></td>
                     <td class="btns"> 
                         <a type="button" class="btn btn-secondary" href="edit.php?id=<?=$id_schet?>">Редактировать</a> 
-                        <a type="button" class="btn btn-secondary <?= disabledDelete($id_schet) ? 'disabled' : '' ?>" href="delete.php?id=<?=$id_schet?>">Удалить</a>
+                        <a type="button" class="btn btn-secondary <?= disabledDelete($id_schet) ? 'disabled' : '' ?>" href="delete.php?id=<?=$id_schet?>" onclick='return confirm(`Вы действительно хотите удалить запись "<?=$number_account?>"?`);'>Удалить</a>
                     </td>
                 </tr>
                 <?php } ?>
